feat!: make transform an interface type
This change introduces a new `Transform` type with a canonical implementation
`TransformFunc` that can be used to transform a string at render time. This
change also updates the `Style` type to use the new `Transform` type instead of
a function type.

We need this to support complex transforms and to be able to compare styles
using the `==` operator.